# release-25.2: storepool: lock individual storepool details instead of all details